PercentageOfLanguages.py
In this part of the code we came up with the following user story. Since it's not convenient enough to find an average number for languages, we decided to find a percentage, which will give an opportunity for user to see what percentage of captions are written in which language.